pyproject.toml
[project]
name = "dkb-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Christoph Dietrich",email = "mail@christoph-dietrich.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
dkb-robo = { path = "./dkb-robo", develop = true }

[tool.poetry]
package-mode = false